refacto(server): delete dbt server class
