+++ server/config.py
@@ -1,0 +1,11 @@
+import os
+from dataclasses import dataclass
+from pathlib import Path
+
+
+@dataclass
+class CONFIG:
+    port: int = os.getenv("PORT", 8000)
+    provider: str = os.getenv("PROVIDER", "GCP")
+    log_level: str = os.getenv("LOG_LEVEL", "INFO")
+    persisted_dir: Path = Path(__file__).parent / "data"
