--- server/lib/storage/__init__.py
@@ -1,5 +1,0 @@
-from server.lib.storage.gcp import GCPStorageBackend
-
-storage_backend = {
-    "GCP": GCPStorageBackend
-}
